changes made

+++ app/controllers/restaurants_controller.rb
@@ -2,21 +2,24 @@ class RestaurantsController < ApplicationController
   rescue_from ActiveRecord::RecordNotFound, with: :render_not_found_response
 
 def index
-  restaurants=Restaurant.all
-  render json: restaurants, include: ['']
+  restaurant =Restaurant.all
+  render json: restaurant, include: ['']
 end
 
 def show
   restaurant = Restaurant.find(params[:id])
-  render json: restaurant, include: ['pizas']
+  render json: restaurant, include: ['pitzas']
 
 
 end
 
 def destroy
-  restaurant = Restaurant.find(params[:id])
-  restaurant.destroy
-  render json: restaurant.piza
+  restaurant = Restaurant.find_by(id: params[:id])
+  if restaurant.
+    restaurant.destroy
+  else
+  render json: {error: "Restaurant not found"}, status: :not_found
+end
 end
 
 private
@@ -24,5 +27,6 @@ private
 def render_not_found_response
 render json: { error: "Restaurant not found" }, status: :not_found
 end
-
 end
+
+
